+++ src/helper/variables.ts
@@ -12,7 +12,7 @@ export const statblockPopoverId = `${ID}/statblock-popover`;
 export const rollLogPopoverId = `${ID}/dice-log`;
 export const rollMessageChannel = `${ID}.roll-event`;
 
-export const version = "2.3.0";
+export const version = "2.3.1";
 
 export const changelogModal: Modal = {
     id: modalId,